+++ saleor/graphql/order/mutations/draft_order_complete.py
@@ -94,7 +94,6 @@ class DraftOrderComplete(BaseMutation):
         order.save()
 
         channel = order.channel
-        channel_slug = channel.slug
         order_lines_info = []
         for line in order.lines.all():
             if line.variant.track_inventory or line.variant.is_preorder_active():
@@ -108,13 +107,13 @@ class DraftOrderComplete(BaseMutation):
                         allocate_stocks(
                             [line_data],
                             country,
-                            channel_slug,
+                            channel,
                             manager,
                             check_reservations=is_reservation_enabled(site.settings),
                         )
                         allocate_preorders(
                             [line_data],
-                            channel_slug,
+                            channel.slug,
                             check_reservations=is_reservation_enabled(site.settings),
                         )
                 except InsufficientStock as exc: